support multiple lua versions, merged v2g libraries
