+++ Installer/Build_Installer.bat
@@ -1,14 +1,11 @@
-:: Script to build DLLs and create Installer. Should be run in MinGW Bash (e.g. git-bash)
+:: Script to create the installer. Make sure the v2gLib is built before.
+:: Should be run in MinGW Bash (e.g. git-bash)
 :: note: Inno Setup must be installed and added to Path
 
 @echo off
-:: create DLLs
+:: copy v2gLib binaries
 cd "%~dp0../V2G_Libraries/"
-call build_all_win.bat
-if %ERRORLEVEL% NEQ 0 goto :failed
-copy "CertificateInfos\bin\v2gX509CertInfos.dll" "%~dp0..\Wireshark\"
-if %ERRORLEVEL% NEQ 0 goto :failed
-copy "V2GDecoder\bin\v2gLuaDecoder.dll" "%~dp0..\Wireshark\"
+copy "v2gLib\bin\v2gLib*.dll" "%~dp0..\Wireshark\"
 if %ERRORLEVEL% NEQ 0 goto :failed
 
 :: prepare installer readme
