--- V2G_Libraries/CertificateInfos/.gitignore
@@ -1,4 +1,0 @@
-.vscode
-bin
-*.o
-*.d
