--- V2G_Libraries/CertificateInfos/main.rc
@@ -1,28 +1,0 @@
-#include <windows.h>
-
-#define VER_FILEVERSION             1,4,1,0
-#define VER_FILEVERSION_STR         "1.4.1.0\0"
-#define VER_COMPANYNAME_STR         "dSPACE GmbH"
-#define VER_PRODUCTNAME_STR         "V2gCertificateInfos"
-#define VER_PRODUCTVERSION          1,4,1,0
-#define VER_PRODUCTVERSION_STR      "1.4.1.0\0"
-
-VS_VERSION_INFO VERSIONINFO
-FILEVERSION VER_FILEVERSION
-PRODUCTVERSION VER_PRODUCTVERSION
-BEGIN
-    BLOCK "StringFileInfo"
-    BEGIN
-        BLOCK "040904b0"
-        BEGIN
-            VALUE "CompanyName",        VER_COMPANYNAME_STR
-            VALUE "FileVersion",        VER_FILEVERSION_STR
-            VALUE "LegalCopyright",     "Copyright 2024, dSPACE GmbH. All Rights Reserved\0"
-            VALUE "ProductVersion",     VER_PRODUCTVERSION_STR
-        END
-    END
-    BLOCK "VarFileInfo"
-    BEGIN
-        VALUE "Translation", 0x409, 1200
-    END
-END
