+++ V2G_Libraries/v2gLib/src/GnuTLSHelper.cpp
@@ -21,22 +21,9 @@
 #include <iomanip>
 
 #include "CertHelper.h"
+#include "Utils.h"
 
 constexpr int MAX_STRING_LEN = 512;
-
-std::string uint8_to_hex_string(const uint8_t *v, const size_t s)
-{
-    std::stringstream ss;
-
-    ss << std::hex << std::setfill('0');
-
-    for (size_t i = 0; i < s; i++)
-    {
-        ss << std::hex << std::setw(2) << static_cast<int>(v[i]);
-    }
-
-    return ss.str();
-}
 
 X509CertInfos get_cert_info(std::string x509_content)
 {
