+++ V2G_Libraries/v2gLib/src/Utils.cpp
