+++ V2G_Libraries/v2gLib/src/XmlValidator.cpp
