+++ update_version.bat
@@ -44,27 +44,8 @@ for /F delims^=^ eol^= %%A in ('%SystemRoot%\System32\findstr.exe /N "^" "%FileN
 ) >>"%TempFile%"
 move /y %TempFile% %FileName%
 
-:: update V2GDecoder resource file
-set "FileName=.\V2G_Libraries\V2GDecoder\main.rc"
-del "%TempFile%" 2>nul
-for /F delims^=^ eol^= %%A in ('%SystemRoot%\System32\findstr.exe /N "^" "%FileName%"') do (
-    set "Line=%%A"
-    setlocal EnableDelayedExpansion
-    if not "!Line:#define VER_PRODUCTVERSION =!" == "!Line!" (
-        echo #define VER_PRODUCTVERSION          %newVersion:.=,%,0
-    ) else if not "!Line:#define VER_PRODUCTVERSION_STR=!" == "!Line!" (
-        echo #define VER_PRODUCTVERSION_STR      "%newVersion%.0\0"
-    ) else if not "!Line:#define VER_FILEVERSION =!" == "!Line!" (
-        echo #define VER_FILEVERSION             %newVersion:.=,%,0
-    ) else if not "!Line:#define VER_FILEVERSION_STR=!" == "!Line!" (
-        echo #define VER_FILEVERSION_STR         "%newVersion%.0\0"
-    ) else echo(!Line:*:=!
-    endlocal
-) >>"%TempFile%"
-move /y %TempFile% %FileName%
-
-:: update CertificateInfo resource file
-set "FileName=.\V2G_Libraries\CertificateInfos\main.rc"
+:: update v2gLib resource file
+set "FileName=.\V2G_Libraries\v2gLib\main.rc"
 del "%TempFile%" 2>nul
 for /F delims^=^ eol^= %%A in ('%SystemRoot%\System32\findstr.exe /N "^" "%FileName%"') do (
     set "Line=%%A"
